+++ common/JackNetManager.h
@@ -43,8 +43,8 @@ namespace Jack
 			struct sockaddr_in fAddr;
 			struct sockaddr_in fMcastAddr;
 			int fSockfd;
-			size_t fNSubProcess;
-			size_t fNetJumpCnt;
+			unsigned int fNSubProcess;
+			unsigned int fNetJumpCnt;
 			bool fRunning;
 
 			jack_client_t* fJackClient;
@@ -75,8 +75,8 @@ namespace Jack
 			void FreePorts();
 			void Exit();
 
-			int Send ( char* buffer, size_t size, int flags );
-			int Recv ( size_t size, int flags );
+			int Send ( char* buffer, unsigned int size, int flags );
+			int Recv ( unsigned int size, int flags );
 			int Process();
 		public:
 			JackNetMaster ( JackNetMasterManager* manager, session_params_t& params, struct sockaddr_in& address, struct sockaddr_in& mcast_addr );
@@ -97,13 +97,13 @@ namespace Jack
 			const char* fMCastIP;
 			pthread_t fManagerThread;
 			master_list_t fMasterList;
-			size_t fGlobalID;
+			unsigned int fGlobalID;
 			bool fRunning;
-			size_t fPort;
+			unsigned int fPort;
 
 			void Run();
 			JackNetMaster* MasterInit ( session_params_t& params, struct sockaddr_in& address, struct sockaddr_in& mcast_addr );
-			master_list_it_t FindMaster ( size_t client_id );
+			master_list_it_t FindMaster ( unsigned int client_id );
 			void KillMaster ( session_params_t* params );
 			void SetSlaveName ( session_params_t& params );
 			int Process();
